// admin/src/app/cadastro/types.ts
export interface Funcionario {
  id: number;
  nome: string;
  departamento: string;
  endereco: string;
  email: string;
}

/*
  "id": 1,
  "nome":"Alonzo Birden",
  "foto":"https://robohash.org/praesentiumomnisdolore.png?size=50x50&set=set1",
  "email":"abirden0@mac.com",
  "subobjeto": {
    "propriedade1": "Valor 1",
    "propriedade2": "Valor 2"
  */